pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "whisper"
version = "0.1.0"
description = "Local transcription CLI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "openai-whisper>=20250625",
  "typer>=0.21.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0.2",
  "ruff>=0.15.0",
]

[project.scripts]
whisp = "whisp.cli:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["whisp"]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
